pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "career-ops"
version = "0.1.0"
description = "AI-powered, RAG-grounded job-search system for ML / GenAI / MLOps roles."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Lokesh Reddy", email = "lokeshsaireddyk@gmail.com" }]
keywords = ["rag", "job-search", "llm", "career", "genai"]

dependencies = [
  # LLM + orchestration
  "anthropic>=0.39",
  "openai>=1.40",
  "langchain>=0.3",
  "langchain-community>=0.3",
  # Embeddings + vector store
  "sentence-transformers>=3.0",
  "faiss-cpu>=1.8",
  # Data / API
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "pyyaml>=6.0",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  # Ingestion / scraping
  "playwright>=1.46",
  "httpx>=0.27",
  "beautifulsoup4>=4.12",
  "markdownify>=0.13",
  # Rendering
  "jinja2>=3.1",
  "weasyprint>=62",
  "markdown-it-py>=3.0",
  # UX / CLI / dashboard
  "typer>=0.12",
  "rich>=13.7",
  "structlog>=24.1",
  "tqdm>=4.66",
  "streamlit>=1.37",
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
  "types-PyYAML",
  "pre-commit>=3.8",
]

[project.scripts]
career-ops = "career_ops.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
